Added user input validation

+++ Products/Application/Filters/Filter.cs
@@ -4,8 +4,8 @@ namespace Products.Application.Filters
 {
     public class Filter
     {
-        public int MinPrice { get; set; } = 0;
-        public int MaxPrice { get; set; } = int.MaxValue;
+        public double MinPrice { get; set; } = 0;
+        public double MaxPrice { get; set; } = int.MaxValue;
         public Guid? SellerId { get; set; } = null;
     }
 }
